fix: fix crash of circularvideobuffer

+++ Holovibes/cuda_includes/moments_treatments.cuh
@@ -14,7 +14,7 @@ void subtract_frame_from_sum(const float* const new_frame,
                              float* const sum_image,
                              cudaStream_t stream);
 
-void compute_mean(float* output, float* input, const int time_window, const size_t frame_size, cudaStream_t stream);
+void compute_mean(float* output, float* input, const size_t time_window, const size_t frame_size, cudaStream_t stream);
 
 void temporal_mean(float* output,
                    float* input,
